Remove legacy Crashlytics onboarding calls
